+++ packages/web/src/components/common/header/search-menu-modal/SearchMenuModal.stories.tsx
