+++ packages/web/src/components/common/header/search-menu-modal/SearchMenuModal.styles.ts
